Add a function to kick off OAuth authentication
Add a function setting up OAuth for Podchaser, which would be required for
mutating queries later on. Right now we are lacking a programmatic way to
obtain the resulting auth token because Podchaser doesn't allow me to set
the API's redirect URI to localhost (which would allow listening for that
request ourselves and acting on it locally).

Signed-off-by: Bjoern Doebel <bjoern.doebel@gmail.com>